restructure frontend files
